back/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# ML & AI
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.25.0
sentencepiece>=0.1.99

# Image processing
Pillow>=10.1.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
num2words>=0.5.13

# Gradio UI
gradio>=4.0.0

# HuggingFace
huggingface-hub>=0.19.0
safetensors>=0.4.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

